tonypedia-backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.31.0
sqlalchemy>=2.0.35
psycopg2-binary>=2.9.12
requests>=2.32.3
python-dotenv>=1.0.1
anthropic>=0.42.0
voyageai>=0.4.0
numpy>=1.26.4
pydantic>=2.8.2
python-multipart>=0.0.6
setuptools>=70.0.0